test(main): 端口解析用例移出 test_main.py 以绕开既有循环导入
tests/test_main.py 因 app.services 的循环导入无法被收集，放在其中的
resolve_http_port 用例从未真正执行。拆到只依赖 main 的独立文件后，
4 个用例与 3 个子用例均可正常运行。
