+++ tests/test_main.py
@@ -59,29 +59,6 @@ class MainDevelopmentEnvironmentTest(unittest.TestCase):
             self.assertFalse(main.is_hosted_launch())
 
 
-class ResolveHttpPortTest(unittest.TestCase):
-    def test_production_keeps_legacy_port(self) -> None:
-        with patch.dict(os.environ, {"AUTO_MAS_HTTP_PORT": ""}):
-            self.assertEqual(main.resolve_http_port(False), main.DEFAULT_HTTP_PORT)
-
-    def test_development_uses_dedicated_port(self) -> None:
-        """开发环境错开端口，用户已装正式版仍可同时运行。"""
-
-        with patch.dict(os.environ, {"AUTO_MAS_HTTP_PORT": ""}):
-            self.assertEqual(main.resolve_http_port(True), main.DEV_HTTP_PORT)
-
-    def test_environment_variable_overrides_both(self) -> None:
-        with patch.dict(os.environ, {"AUTO_MAS_HTTP_PORT": "40000"}):
-            self.assertEqual(main.resolve_http_port(True), 40000)
-            self.assertEqual(main.resolve_http_port(False), 40000)
-
-    def test_invalid_environment_variable_falls_back(self) -> None:
-        for raw in ("0", "70000", "abc"):
-            with self.subTest(raw=raw):
-                with patch.dict(os.environ, {"AUTO_MAS_HTTP_PORT": raw}):
-                    self.assertEqual(main.resolve_http_port(False), main.DEFAULT_HTTP_PORT)
-
-
 class TelemetryDevelopmentTest(unittest.TestCase):
     def setUp(self) -> None:
         telemetry._sentry_release = None
